ekf2: move class definition to separate header
